POC for using nexthop groups in Submariner
This is a component of the multi-active gateway support. It is not a
complete implementation in that it doesn't handle all cases.

A major challenge is that the Submariner code uses vishvananda/netlink for
route network configuration, but the vishnananda/netlink package doesn't
support nexthop groups. This makes using nexthop groups in Submariner very
messy and error prone.

After doing this work, I'm thinking that it may be better to start
the POC with the per-route "nexthop" multi-path option that is currently
supported by vishnananda/netlink. Then, if everything works out, add
nexthop support to vishnananda/netlink and then Submariner.

Signed-off-by: Andre Fredette <afredette@redhat.com>